[MonorepoBuilder] Simplify Autoload merging
Closes #1203
