--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/Source/composer.json
@@ -1,17 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\": "src",
-            "Shopsys\\": [
-                "app/",
-                "src/Shopsys/"
-            ]
-        },
-        "files": [
-            "src/SomeFile.php"
-        ],
-        "classmap": [
-            "src/SomeClass.php"
-        ]
-    }
-}
