--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/SourceIdenticalNamespaces/PackageA/composer.json
@@ -1,8 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\Core\\": "src/core",
-            "App\\Shared\\": "src/shared"
-        }
-    }
-}
