--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/SourceIdenticalNamespaces/PackageB/composer.json
@@ -1,9 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\Core\\": "src/core-extension",
-            "App\\Model\\": "src/interfaces",
-            "App\\Shared\\": "src/shared"
-        }
-    }
-}
