--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/SourceIdenticalNamespaces/SubA/PackageC/composer.json
@@ -1,8 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\Model\\": "src/models",
-            "App\\Sub\\": "src/package-c"
-        }
-    }
-}
