--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/SourceIdenticalNamespaces/SubB/PackageD/composer.json
@@ -1,7 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\Sub\\": "src/package-d"
-        }
-    }
-}
