--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/SourceOverlappingNamespaces/PackageA/composer.json
@@ -1,11 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\": "src",
-            "Shopsys\\": [
-                "app/",
-                "src/Shopsys/"
-            ]
-        }
-    }
-}
