--- packages/MonorepoBuilder/tests/ComposerJsonDecorator/AutoloadRelativePathComposerJsonDecorator/SourceOverlappingNamespaces/PackageB/composer.json
@@ -1,11 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "App\\PackageB\\": "src",
-            "Shopsys\\PackageB\\": [
-                "app/",
-                "src/Shopsys/"
-            ]
-        }
-    }
-}
