+++ packages/MonorepoBuilder/tests/Package/CombineStringsToArrayJsonMergerTest.php
@@ -4,40 +4,37 @@ namespace Symplify\MonorepoBuilder\Tests\Package;
 
 final class CombineStringsToArrayJsonMergerTest extends AbstractMergeTestCase
 {
-    public function testSharedNamespaces(): void
+    public function testIdenticalNamespaces(): void
     {
         $expectedJson = [
             'autoload' => [
                 'psr-4' => [
-                    'ACME\Another\\' => 'packages/A',
-                    'ACME\Model\Core\\' => ['packages/A', 'packages/B'],
-                    'ACME\\YetAnother\\' => ['packages/A'],
-                    'ACME\\YetYetAnother\\' => 'packages/A',
+                    'App\\Core\\' => [
+                        $this->getRelativeSourcePath() . 'PackageA/src/core',
+                        $this->getRelativeSourcePath() . 'PackageB/src/core-extension'],
+                    'App\\Model\\' => [
+                        $this->getRelativeSourcePath() . 'PackageB/src/interfaces',
+                        $this->getRelativeSourcePath() . 'SubA/PackageC/src/models'
+                    ],
+                    'App\\Shared\\' => [
+                        $this->getRelativeSourcePath() . 'PackageA/src/shared',
+                        $this->getRelativeSourcePath() . 'PackageB/src/shared',
+                    ],
+                    'App\\Sub\\' => [
+                        $this->getRelativeSourcePath() . 'SubA/PackageC/src/package-c',
+                        $this->getRelativeSourcePath() . 'SubB/PackageD/src/package-d'
+                    ],
                 ],
             ],
         ];
 
-        $this->doTestDirectoryMergeToFile(__DIR__ . '/SourceAutoloadSharedNamespaces', $expectedJson);
+        $this->doTestDirectoryMergeToFile(__DIR__ . '/SourceIdenticalNamespaces', $expectedJson);
     }
 
-    public function testIdenticalNamespaces(): void
+    private function getRelativeSourcePath(): string
     {
-        // TODO: define expected behaviour; currently everything we know is: this ain't of any use
-        $NOTexpectedJson = [
-            'autoload' => [
-                'psr-4' => [
-                    'App\\Core\\' => ['src/core', 'src/core-extension'],
-                    'App\\Model\\' => ['src/interfaces', 'src/models'],
-                    'App\\Shared\\' => 'src/shared',
-                    'App\\Sub\\' => ['src/package-c', 'src/package-d'],
-                ],
-            ],
-        ];
+        $prefix = defined('SYMPLIFY_MONOREPO') ? 'packages/MonorepoBuilder/' : '';
 
-        $merged = $this->packageComposerJsonMerger->mergeFileInfos(
-            $this->getFileInfosFromDirectory(__DIR__ . '/SourceIdenticalNamespaces')
-        );
-
-        $this->assertNotSame($NOTexpectedJson, $merged);
+        return $prefix . 'tests/Package/SourceIdenticalNamespaces/';
     }
 }
