--- packages/MonorepoBuilder/tests/Package/SourceAutoloadSharedNamespaces/package-two.json
@@ -1,10 +1,0 @@
-{
-    "autoload": {
-        "psr-4": {
-            "ACME\\Model\\Core\\": "packages/B",
-            "ACME\\Another\\": "packages/A",
-            "ACME\\YetAnother\\": "packages/A",
-            "ACME\\YetYetAnother\\": ["packages/A"]
-        }
-    }
-}
